requirements.txt
django>=2.1.3,<2.2.0
djangorestframework>=3.9.0,<3.10.0
django-rest-knox>=4.0.0,<4.1.0
psycopg2>=2.7.5,<2.8.0
flake8>=3.6.0,<3.7.0
pylint>=2.3.0,<2.4.0
pylint-django>=2.0.0,<2.1.0
